+++ ntex-io/src/io.rs
@@ -1222,7 +1222,7 @@ mod tests {
         assert!(!io.is_wr_backpressure());
 
         // == Enable backpressure
-        ctx.update_read_status(
+        ctx.release_read_buf(
             BytesMut::copy_from_slice(b"1234567890"),
             Poll::Ready(Ok(10)),
         );
@@ -1279,7 +1279,7 @@ mod tests {
         lazy(|cx| io.poll_dispatch(cx)).await;
 
         // == Enable backpressure, 4 bytes in buffer + 4 more
-        ctx.update_read_status(BytesMut::copy_from_slice(b"1234"), Poll::Ready(Ok(4)));
+        ctx.release_read_buf(BytesMut::copy_from_slice(b"1234"), Poll::Ready(Ok(4)));
 
         // dispatcher is woken
         assert!(!io.st().dispatch_task.is_set());
@@ -1301,7 +1301,7 @@ mod tests {
         lazy(|cx| io.poll_dispatch(cx)).await;
 
         // == No backpressure, 4 bytes in buffer + 3 more
-        ctx.update_read_status(BytesMut::copy_from_slice(b"567"), Poll::Ready(Ok(3)));
+        ctx.release_read_buf(BytesMut::copy_from_slice(b"567"), Poll::Ready(Ok(3)));
 
         // read task is paused
         assert!(!io.st().flags.is_read_paused());
@@ -1349,7 +1349,7 @@ mod tests {
         let ctx = IoContext::new(io.get_ref());
 
         // incoming bytes
-        ctx.update_read_status(BytesMut::copy_from_slice(b"1"), Poll::Ready(Ok(1)));
+        ctx.release_read_buf(BytesMut::copy_from_slice(b"1"), Poll::Ready(Ok(1)));
 
         assert!(!io.st().dispatch_task.is_set());
         // rd buffer is ready
@@ -1377,7 +1377,7 @@ mod tests {
         );
 
         // == enable packpressure
-        ctx.update_read_status(BytesMut::copy_from_slice(b"2345678"), Poll::Ready(Ok(7)));
+        ctx.release_read_buf(BytesMut::copy_from_slice(b"2345678"), Poll::Ready(Ok(7)));
         // read backpressure is enabled
         assert!(io.st().flags.is_rd_backpressure());
 
@@ -1408,7 +1408,7 @@ mod tests {
         );
 
         // incoming bytes
-        ctx.update_read_status(BytesMut::copy_from_slice(b"1"), Poll::Ready(Ok(1)));
+        ctx.release_read_buf(BytesMut::copy_from_slice(b"1"), Poll::Ready(Ok(1)));
         assert!(!io.st().dispatch_task.is_set());
         // rd buffer is ready
         assert!(io.st().flags.is_read_ready());