+++ ntex-net/src/compio/io.rs
@@ -126,7 +126,7 @@ async fn read<T>(io: T, ctx: &IoContext)
 where
     T: AsyncRead + AsyncWrite + Clone + Unpin,
 {
-    let mut read_fut = Some(Box::pin(read_buf(&io, ctx.get_read_buf())));
+    let mut read_fut = Some(Box::pin(read_buf(&io, ctx.take_read_buf())));
 
     loop {
         if read_ready(ctx).await {
@@ -135,10 +135,10 @@ where
 
         match select(read_fut.as_mut().unwrap(), not_read_ready(ctx)).await {
             Either::Left(BufResult(result, cbuf)) => {
-                if ctx.update_read_status(cbuf.0, Poll::Ready(result)) == IoTaskStatus::Stop {
+                if ctx.release_read_buf(cbuf.0, Poll::Ready(result)) == IoTaskStatus::Stop {
                     break;
                 }
-                read_fut = Some(Box::pin(read_buf(&io, ctx.get_read_buf())));
+                read_fut = Some(Box::pin(read_buf(&io, ctx.take_read_buf())));
             }
             Either::Right(true) => break,
             Either::Right(false) => (),